♻️[Refactor] router 수정 #21
