--- src/app/admin/block/block.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import Page from "@app/admin/block/image/page";
-import Page from "@app/admin/block/video/page";
-
-const Block = () => {
-  return (
-    <div className="w-full">
-      <Page />
-      <Page />
-    </div>
-  );
-};
-
-export default Block;
