+++ src/app/admin/block/image/page.tsx
@@ -3,8 +3,8 @@ import React, { useRef, useState } from "react";
 import BlockModal from "@app/admin/block/components/block-modal";
 import TextInputBox from "@app/admin/block/components/text-input-box";
 import Image from "next/image";
-import AddButton from "@app/admin/block/components/add-button";
-import ButtonBox from "@app/admin/block/components/button-box";
+import AddButton from "@app/admin/block/components/buttons/add-button";
+import ButtonBox from "@app/admin/block/components/buttons/button-box";
 import ErrorBoundary from "@app/(intro)/components/error-boundary";
 import ImageBox from "@app/admin/block/image/components/image-box";
 import BoundaryImageBox from "@app/admin/block/image/components/image-box";
