+++ src/app/admin/block/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Page = () => {
+  return <div>block page</div>;
+};
+
+export default Page;
